Moved the file to its correct position
